score added, paddle boundaries
im seeing a bug that i know the source of but not how to fix: if the paddle is near the upper or lower edge and stops, the next move will take it past the edge. this is because the unit of movement is greater than one... i think i just thought of a fix. nvm